requirements.txt
streamlit>=1.20.0
torch>=1.13.0
torchvision>=0.14.0
openai>=1.0.0
Pillow>=9.0.0
numpy